some changes

+++ Warehouse/Goods/Food.cs
@@ -4,7 +4,7 @@
     {
         public DateTime ExpiryDate { get; set; }
 
-        public Food(string category, string nameOfGood, string unitOfMeasure, int unitPrice, int amount, DateTime expiryDate, DateTime dateOfLastDelivery)
+        public Food(string category, string nameOfGood, string unitOfMeasure, double unitPrice, int amount, DateTime expiryDate, DateTime dateOfLastDelivery)
             : base(category, nameOfGood, unitOfMeasure, unitPrice, amount, dateOfLastDelivery)
         {
             ExpiryDate = expiryDate;
